apps/API/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pillow>=10.0.0
openai>=1.0.0

